chore: rename folder tests
