+++ tests/config-tests/smoldot.ts
