+++ tests/interaction-tests/test-proxy-manager.ts
